+++ lang/library/deu.php
@@ -1,16 +1,20 @@
 <?php
 
-$p = array(
+$phrasebook = array(
 
 	'pbook_meta' => array(
 		'name'		=> 'deutsch',
 		'iso6391'	=> 'de',
 		'iso6393'	=> 'deu',
-		'dir'		=> 'ltr',
+		'direction'	=> 'ltr',
 	),
 
-	'fr_title' => 'Titel text',
-	'fr_para'	=> '<p>Dies ist eine kurze, einfache Text, der geschrieben sind mit der Sprache-Modul zu helfen.</p>',
+'fr_title' => 'Titel text',
+'fr_para'	=> 'Dies ist eine kurze, einfache Text, der geschrieben sind mit der Sprache-Modul zu helfen.',
+'name'		=> 'Name der Sprache',
+'iso6391'	=> 'ISO639-1',
+'iso6393'	=> 'ISO639-3',
+'direction'	=> 'Text Richtung',
 
 );