+++ lang/library/eng.php
@@ -1,16 +1,20 @@
 <?php
 
-$p = array(
+$phrasebook = array(
 
 	'pbook_meta' => array(
 		'name'		=> 'english',
 		'iso6391'	=> 'en',
 		'iso6393'	=> 'eng',
-		'dir'		=> 'ltr',
+		'direction'	=> 'ltr',
 	),
 
-	'fr_title' => 'Title text',
-	'fr_para'	=> '<p>This is a short, simple text that is written to help with the language module.</p>',
+'fr_title' => 'Title text',
+'fr_para'	=> 'This is a short, simple text that is written to help with the language module.',
+'name'		=> 'Language name',
+'iso6391'	=> 'ISO639-1',
+'iso6393'	=> 'ISO639-3',
+'direction'	=> 'Text direction',
 
 );