+++ lang/library/hun.php
@@ -1,16 +1,20 @@
 <?php
 
-$p = array(
+$phrasebook = array(
 
 	'pbook_meta' => array(
 		'name'		=> 'magyar',
 		'iso6391'	=> 'hu',
 		'iso6393'	=> 'hun',
-		'dir'		=> 'ltr',
+		'direction'	=> 'ltr',
 	),
 
-	'fr_title' => 'Cím szöveg',
-	'fr_para'	=> '<p>Ez egy rövid, egyszerű szöveg van írva, hogy segítsen a nyelvi modul.</p>',
+'fr_title' => 'Cím szöveg',
+'fr_para'	=> 'Ez egy rövid, egyszerű szöveg van írva, hogy segítsen a nyelvi modul.',
+'name'		=> 'Nyelv neve',
+'iso6391'	=> 'ISO639-1',
+'iso6393'	=> 'ISO639-3',
+'direction'	=> 'Szöveg irány',
 
 );