+++ lang/library/jpn.php
@@ -1,16 +1,20 @@
 <?php
 
-$p = array(
+$phrasebook = array(
 
 	'pbook_meta' => array(
 		'name'		=> '日本語',
 		'iso6391'	=> 'ja',
 		'iso6393'	=> 'jpn',
-		'dir'		=> 'ltr',
+		'direction'	=> 'ltr',
 	),
 
-	'fr_title' => 'タイトルのテキスト',
-	'fr_para'	=> '<p>これは、言語モジュールを助けるために書かれている短い、単純なテキストである。</p>',
+'fr_title' => 'タイトルのテキスト',
+'fr_para'	=> 'これは、言語モジュールを助けるために書かれている短い、単純なテキストである。',
+'name'		=> '言語名',
+'iso6391'	=> 'ISO639-1',
+'iso6393'	=> 'ISO639-3',
+'direction'	=> 'テキスト方向',
 
 );