+++ lang/library/tlh.php
@@ -1,16 +1,20 @@
 <?php
 
-$p = array(
+$phrasebook = array(
 
 	'pbook_meta' => array(
 		'name'		=> 'tlhIngan Hol',
 		'iso6391'	=> 'qh',	// --- Locally used code. There is no registered ISO639-1 code for klingon.
 		'iso6393'	=> 'tlh',
-		'dir'		=> 'ltr',
+		'direction'	=> 'ltr',
 	),
 
-	'fr_title' => 'title bIngDaq ghItlh leghlu\'',
-	'fr_para'	=> '<p>\'oH bIngDaq ghItlh ngaj, nap leghlu\' \'e\' QaH Hol bobcho\' ghItlh.</p>',
+'fr_title' => 'title bIngDaq ghItlh leghlu\'',
+'fr_para'	=> '\'oH bIngDaq ghItlh ngaj, nap leghlu\' \'e\' QaH Hol bobcho\' ghItlh.',
+'name'		=> 'Hol pong',
+'iso6391'	=> 'ISO639-1',
+'iso6393'	=> 'ISO639-3',
+'direction'	=> 'bIngDaq ghItlh leghlu\' lurgh',
 
 );