+++ lang/settings.php
@@ -8,10 +8,10 @@
 // ============================================================================
 
 $settings = array(
-	'default'	=> 'swe',			// --- Default language to use if no value is passed to the constructor. This language file must be present in the library and should be complete.
-	'nophrase'	=> '(( -- ERROR: Phrase not in phrasebook. -- ))',		// --- String to replace any non-existing phrase.
-	'getvar'	=> 'lang',			// --- Name of $_GET-variable for the language.
-	'll_class'	=> 'lang_list',		// --- Class name of the language list.
+	'default_lang'		=> 'swe',			// --- Default language to use if no value is passed to the constructor. This language file must be present in the library and should be complete.
+	'no_phrase'			=> '(( -- ERROR: Phrase not in phrasebook. -- ))',		// --- String to replace any non-existing phrase.
+	'lang_variable'		=> 'lang',			// --- Name of $_GET-variable for the language.
+	'lang_list_class'	=> 'languagelist',		// --- Class name of the language list.
 
 
 
@@ -19,18 +19,19 @@ $settings = array(
 // There shouldn't be any need to edit the following stuff at all.
 
 // --- Settings for file- and directory names.
-	'imgdir'	=> 'images',		// --- Directory where the images are located. Must be a child directory of the one this file resides in.
-	'pbdir'		=> 'library',		// --- Directory where the phrasebooks are located. Must be a child directory of the one this file resides in.
-	'libfile' 	=> 'library.json',	// --- Name of the file that stores the static library info.
-	'chgfile' 	=> '.changefile',	// --- Name of the file used to check for changes in the library.
+	'imagedir'		=> 'images',		// --- Directory where the images are located. Must be a child directory of the one this file resides in.
+	'phrasebookdir'	=> 'library',		// --- Directory where the phrasebooks are located. Must be a child directory of the one this file resides in.
+	'libraryfile' 	=> 'library.json',	// --- Name of the file that stores the static library info.
+	'changefile' 	=> '.changefile',	// --- Name of the file used to check for changes in the library.
 );
 
 // --- A few modifications to the settings to make sure all the paths work as intended.
-$installdir = dirname(__FILE__).'/';
-$settings['dirurl'] = $_SERVER['DOMAIN_NAME'].dirname(str_replace($_SERVER['DOCUMENT_ROOT'], '', __FILE__));
-$settings['imgdir'] = $settings['dirurl'].'/'.$settings['imgdir'];
-$settings['pbdir'] = $installdir.$settings['pbdir'];
-$settings['libfile'] = $installdir.$settings['libfile'];
-$settings['chgfile'] = $installdir.$settings['chgfile'];
+$installdir					= dirname(__FILE__).'/';
+$settings['dirurl']			= 'http://'.$_SERVER['DOMAIN_NAME'].dirname(str_replace($_SERVER['DOCUMENT_ROOT'], '', __FILE__));
+$settings['imagedirurl']	= $settings['dirurl'].'/'.$settings['imagedir'];
+$settings['imagedir']		= $installdir.$settings['imagedir'];
+$settings['phrasebookdir']	= $installdir.$settings['phrasebookdir'];
+$settings['libraryfile']	= $installdir.$settings['libraryfile'];
+$settings['changefile']		= $installdir.$settings['changefile'];
 
 ?>